4.21

+++ viikko4/blogilista/controllers/blogs.js
@@ -37,8 +37,23 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+  if (!user) {
+    return response.status(400).json({ error: 'userId missing or invalid' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndDelete(request.params.id)
+    await user.updateOne({ $pull: {blogs: request.params.id}})
+    return response.status(204).end()
+  } 
+  return response.status(400).json({ error: 'not authorized to delete this blog' })
+
 })
 
 blogsRouter.patch('/:id', async (request, response) => {
